LPS-52484 This is the new pattern in modules
