+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-query/bnd.bnd
@@ -1,7 +1,7 @@
 Bundle-Name: Liferay Dynamic Data Mapping Query
-Bundle-SymbolicName: com.liferay.dynamicdatamapping.query
+Bundle-SymbolicName: com.liferay.dynamic.data.mapping.query
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.dynamicdatamapping.query.api
+Export-Package: com.liferay.dynamic.data.mapping.query.api
 Import-Package:\
 	!javax.imageio.*,\
 	!javax.print.*,\
