+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-query/src/com/liferay/dynamic/data/mapping/query/impl/DDMFormValuesQueryListener.java
@@ -12,27 +12,27 @@
  * details.
  */
 
-package com.liferay.dynamicdatamapping.query.impl;
+package com.liferay.dynamic.data.mapping.query.impl;
 
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueMatchesAllMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueMatchesAnyMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueNameMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueTypeMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormFieldValueValueMatcher;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormValuesFilter;
-import com.liferay.dynamicdatamapping.query.impl.model.DDMFormValuesFilterImpl;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryBaseListener;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.AttributeTypeContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.AttributeValueContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.FieldSelectorContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.FieldSelectorExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.LocaleExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.PredicateAndExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.PredicateEqualityExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.PredicateOrExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.SelectorExpressionContext;
-import com.liferay.dynamicdatamapping.query.impl.parser.DDMFormValuesQueryParser.StepTypeContext;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueMatchesAllMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueMatchesAnyMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueNameMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueTypeMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormFieldValueValueMatcher;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormValuesFilter;
+import com.liferay.dynamic.data.mapping.query.impl.model.DDMFormValuesFilterImpl;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryBaseListener;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.AttributeTypeContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.AttributeValueContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.FieldSelectorContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.FieldSelectorExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.LocaleExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.PredicateAndExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.PredicateEqualityExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.PredicateOrExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.SelectorExpressionContext;
+import com.liferay.dynamic.data.mapping.query.impl.parser.DDMFormValuesQueryParser.StepTypeContext;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
