+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-query/src/com/liferay/dynamic/data/mapping/query/impl/parser/DDMFormValuesQueryListener.java
@@ -14,7 +14,7 @@
  * details.
  */
 
-package com.liferay.dynamicdatamapping.query.impl.parser;
+package com.liferay.dynamic.data.mapping.query.impl.parser;
 
 import org.antlr.v4.runtime.misc.NotNull;
 import org.antlr.v4.runtime.tree.ParseTreeListener;
